refactor: 역할 분리 및 로직 수정 - AuthService가 로그인, 회원가입, 로그아웃, 탈퇴 등 인증 관련 요청 처리하도록 수정 - 소셜 플랫폼과 관련된 로직은 SocialPlatformService로 분리 - 그 외 로직 개선
