+++ src/main/java/com/dongyang/dongpo/domain/auth/service/AuthService.java
@@ -1,40 +1,18 @@
 package com.dongyang.dongpo.domain.auth.service;
 
 
+import com.dongyang.dongpo.common.auth.jwt.JwtService;
 import com.dongyang.dongpo.common.exception.CustomException;
 import com.dongyang.dongpo.common.exception.ErrorCode;
-import com.dongyang.dongpo.common.util.auth.AppleKeyGenerator;
-import com.dongyang.dongpo.common.util.auth.ApplePublicKeyGenerator;
-import com.dongyang.dongpo.common.util.auth.ApplePublicKeyResponse;
-import com.dongyang.dongpo.common.util.auth.AppleTokenParser;
 import com.dongyang.dongpo.domain.auth.dto.*;
-import com.dongyang.dongpo.domain.auth.entity.AppleRefreshToken;
-import com.dongyang.dongpo.domain.auth.repository.AppleRefreshTokenRepository;
 import com.dongyang.dongpo.domain.member.entity.Member;
-import com.dongyang.dongpo.domain.member.entity.Member.SocialType;
 import com.dongyang.dongpo.domain.member.service.MemberService;
 import io.jsonwebtoken.Claims;
+import io.micrometer.common.util.StringUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.json.JSONArray;
-import org.json.JSONObject;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.reactive.function.BodyInserters;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.reactive.function.client.WebClientResponseException;
-import org.springframework.web.util.UriBuilder;
-
-import java.security.PublicKey;
-import java.util.Date;
-import java.util.Map;
-import java.util.Objects;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 @Service
 @Slf4j
@@ -42,355 +20,125 @@ import java.util.stream.IntStream;
 @RequiredArgsConstructor
 public class AuthService {
 
+    private final JwtService jwtService;
     private final MemberService memberService;
-    private final AppleTokenParser appleTokenParser;
-    private final AppleKeyGenerator appleKeyGenerator;
-    private final ApplePublicKeyGenerator applePublicKeyGenerator;
-    private final AppleRefreshTokenRepository appleRefreshTokenRepository;
+    private final SocialPlatformService socialPlatformService;
 
-    @Value("${kakao.client_id}")
-    private String clientId;
+    // 소셜 로그인 처리
+    public JwtToken handleLogin(final String oauthAccessToken) {
+        final UserInfo userInfo = getKakaoUserInfo(oauthAccessToken);
 
-    @Value("${kakao.redirect_uri}")
-    private String redirectUri;
+        // 존재하는 회원일 경우 엔티티 조회 / 존재하지 않는 회원일 경우 회원가입 진행
+        Member member = memberService.validateMemberExistence(userInfo.getEmail(), userInfo.getId()) ?
+                memberService.findByEmail(userInfo.getEmail()) : memberService.registerNewMember(userInfo);
 
-    @Value("${kakao.terms_tag.service_terms}")
-    private String serviceTermsTag;
+        // 로그인 처리
+        return jwtService.createTokenForLoginMember(member);
+    }
 
-    @Value("${kakao.terms_tag.marketing_terms}")
-    private String marketingTermsTag;
+    // 애플 소셜 로그인 처리
+    public AppleLoginResponse handleAppleLogin(final AppleLoginDto appleLoginDto) {
+        final Claims appleUserClaims = getAppleUserInfo(appleLoginDto);
+        final String socialId = appleUserClaims.get("sub", String.class);
+        final String email = appleUserClaims.get("email", String.class);
 
-    @Value("${apple.client.id}")
-    private String appleClientId;
+        // 한 번 탈퇴한 애플 사용자는 인증 서버로부터 이메일이 반환되지 않음
+        if (StringUtils.isBlank(email))
+            throw new CustomException(ErrorCode.MEMBER_ALREADY_LEFT);
 
-    private final String appleAuthUrl = "https://appleid.apple.com";
+        // 이미 존재하는 회원일 경우 로그인 처리
+        if (memberService.validateMemberExistence(email, socialId)) {
+            final Member member = memberService.findByEmail(email);
+            return AppleLoginResponse.responseJwtToken(jwtService.createTokenForLoginMember(member));
+        }
 
-    public JwtToken kakaoCallback(String AccessCode) {
-        WebClient webClient = WebClient.builder()
-                .baseUrl("https://kauth.kakao.com/oauth/token")
-                .defaultHeader("Content-Type", "application/x-www-form-urlencoded;charset=utf-8")
+        // 존재하지 않는 회원인 경우 필수 클레임 정보 반환
+        return AppleLoginResponse.responseClaims(appleUserClaims);
+    }
+
+    // 애플 회원 가입 수행 메소드(추가 정보 기입)
+    public JwtToken continueAppleSignup(final AppleSignupContinueDto appleSignupContinueDto) {
+        // 서비스 이용 약관에 동의 하지 않았을 경우 가입 불허
+        if (!appleSignupContinueDto.getIsMarketingTermsAgreed())
+            throw new CustomException(ErrorCode.SERVICE_TERMS_NOT_AGREED);
+
+        // 이미 가입된 회원인지, 가입 가능한 회원인지 검증
+        if (memberService.validateMemberExistence(appleSignupContinueDto.getEmail(), appleSignupContinueDto.getSocialId()))
+            throw new CustomException(ErrorCode.MEMBER_ALREADY_EXISTS);
+
+        // 회원 정보 생성
+        final UserInfo userInfo = UserInfo.builder()
+                .email(appleSignupContinueDto.getEmail())
+                .nickname(appleSignupContinueDto.getNickname())
+                .birthyear(appleSignupContinueDto.getBirthday().substring(0, 4))
+                .birthday(appleSignupContinueDto.getBirthday().substring(5))
+                .gender(appleSignupContinueDto.getGender())
+                .provider(Member.SocialType.APPLE)
+                .id(appleSignupContinueDto.getSocialId())
                 .build();
 
-        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
-        params.add("grant_type", "authorization_code");
-        params.add("client_id", clientId);
-        params.add("redirect_uri", redirectUri);
-        params.add("code", AccessCode);
-
-        String responseBody = webClient.post()
-                .uri(UriBuilder::build)
-                .body(BodyInserters.fromFormData(params))
-                .retrieve()
-                .bodyToMono(String.class)
-                .block();
-
-        JSONObject jsonObject = new JSONObject(responseBody);
-        String accessToken = jsonObject.getString("access_token");
-        return getKakaoUserInfo(accessToken);
+        // 존재하지 않는 회원인 경우 회원가입 진행
+        final Member member = memberService.registerNewMember(userInfo);
+        return jwtService.createTokenForLoginMember(member);
     }
 
-    public JwtToken getKakaoUserInfo(String accessToken) {
-        try {
-            WebClient webClient = WebClient.builder()
-                    .baseUrl("https://kapi.kakao.com/v2/user/me")
-                    .defaultHeader("Content-Type", "application/x-www-form-urlencoded;charset=utf-8")
-                    .defaultHeader("Authorization", "Bearer " + accessToken)
-                    .build();
-
-            String responseBody = webClient.post()
-                    .retrieve()
-                    .bodyToMono(String.class)
-                    .block();
-
-            JSONObject jsonObject = new JSONObject(responseBody);
-            JSONObject kakaoAccount = jsonObject.getJSONObject("kakao_account");
-            JSONObject profile = kakaoAccount.getJSONObject("profile");
-
-            String email = kakaoAccount.getString("email");
-            String id = String.valueOf(jsonObject.getLong("id"));
-            String birthyear = kakaoAccount.getString("birthyear");
-            String birthday = kakaoAccount.getString("birthday").substring(0, 2) + "-" + kakaoAccount.getString("birthday").substring(2, 4);
-            String nickname = profile.getString("nickname");
-            Member.Gender gender;
-
-            if (kakaoAccount.getString("gender").equals("female"))
-                gender = Member.Gender.GEN_FEMALE;
-            else if (kakaoAccount.getString("gender").equals("male"))
-                gender = Member.Gender.GEN_MALE;
-            else
-                gender = Member.Gender.NONE;
-
-            // 간편 가입 과정에서 동의한 서비스 이용 약관 조회
-            Map<String, Boolean> termsAgreementInfo = getServiceTermsAgreementInfoFromKakao(accessToken);
-
-            return memberService.socialSave(UserInfo.builder()
-                    .id(id)
-                    .email(email)
-                    .nickname(nickname)
-                    .birthyear(birthyear)
-                    .birthday(birthday)
-                    .gender(gender)
-                    .provider(SocialType.KAKAO)
-                    .isServiceTermsAgreed(termsAgreementInfo.get(serviceTermsTag))
-                    .isMarketingTermsAgreed(termsAgreementInfo.get(marketingTermsTag))
-                    .build());
-        } catch (WebClientResponseException e) {
-            if (e.getStatusCode().value() == 401)
-                throw new CustomException(ErrorCode.SOCIAL_TOKEN_NOT_VALID);
-            else
-                throw e;
-        }
+    // 카카오 로그인 콜백 처리
+    public JwtToken kakaoCallback(final String accessCode) {
+        return handleLogin(socialPlatformService.kakaoCallback(accessCode));
     }
 
-    private Map<String, Boolean> getServiceTermsAgreementInfoFromKakao(String accessToken) {
-        try {
-            WebClient webClient = WebClient.builder()
-                    .baseUrl("https://kapi.kakao.com/v2/user/service_terms")
-                    .defaultHeader("Content-Type", "application/x-www-form-urlencoded;charset=utf-8")
-                    .defaultHeader("Authorization", "Bearer " + accessToken)
-
-                    .build();
-
-            String responseBody = webClient.get()
-                    .uri(uriBuilder -> uriBuilder
-                            .queryParam("result", "app_service_terms")
-                            .build())
-                    .retrieve()
-                    .bodyToMono(String.class)
-                    .block();
-
-            JSONArray serviceTerms = new JSONObject(responseBody).getJSONArray("service_terms");
-            return IntStream.range(0, serviceTerms.length())
-                    .mapToObj(serviceTerms::getJSONObject)
-                    .collect(Collectors.toMap(
-                            term -> term.getString("tag"),
-                            term -> term.getBoolean("agreed")
-                    ));
-        } catch (WebClientResponseException e) {
-            if (e.getStatusCode().value() == 401)
-                throw new CustomException(ErrorCode.SOCIAL_TOKEN_NOT_VALID);
-            else
-                throw e;
-        }
+    // 카카오 사용자 정보 조회
+    private UserInfo getKakaoUserInfo(final String accessToken) {
+        return socialPlatformService.getKakaoUserInfo(accessToken);
     }
 
-    /*
-    public JwtToken getNaverUserInfo(String accessToken) {
-        try {
-            WebClient webClient = WebClient.builder()
-                    .baseUrl("https://openapi.naver.com/v1/nid/me")
-                    .defaultHeader("Content-Type", "application/x-www-form-urlencoded;charset=utf-8")
-                    .defaultHeader("Authorization", "Bearer " + accessToken)
-                    .build();
-
-            String responseBody = webClient.post()
-                    .retrieve()
-                    .bodyToMono(String.class)
-                    .block();
-
-            JSONObject jsonObject = new JSONObject(responseBody);
-            JSONObject response = jsonObject.getJSONObject("response");
-
-            String email = response.getString("email");
-            String birthyear = response.getString("birthyear");
-            String birthday = response.getString("birthday");
-            String id = response.getString("id");
-            String nickname = response.getString("nickname");
-            Member.Gender gender;
-
-            if (response.getString("gender").equals("F"))
-                gender = Member.Gender.GEN_FEMALE;
-            else if (response.getString("gender").equals("M"))
-                gender = Member.Gender.GEN_MALE;
-            else
-                gender = Member.Gender.NONE;
-
-            return memberService.socialSave(UserInfo.builder()
-                    .id(id)
-                    .email(email)
-                    .nickname(nickname)
-                    .gender(gender)
-                    .birthyear(birthyear)
-                    .birthday(birthday)
-                    .provider(SocialType.NAVER)
-                    .build());
-        } catch (WebClientResponseException e) {
-            if (e.getStatusCode().value() == 401)
-                throw new CustomException(ErrorCode.SOCIAL_TOKEN_NOT_VALID);
-            else
-                throw e;
-        }
-    }
-     */
-
-    public void doLogout(Member member) {
-        memberService.handleLogout(member);
+    // 애플 사용자 정보 조회
+    private Claims getAppleUserInfo(final AppleLoginDto appleLoginDto) {
+        return socialPlatformService.getAppleUserInfo(appleLoginDto);
     }
 
-    public void doLeave(Member member) {
+    // 로그아웃
+    public void logout(final Member member) {
+        jwtService.expireExistingTokens(member.getEmail());
+    }
+
+    // 탈퇴
+    public void withdraw(final Member member) {
         // 추후 소셜 로그인 플랫폼 추가 시 수정
         switch (member.getSocialType()) {
             case APPLE:
-                revokeToken(member);
+                handleAppleLoginWithdrawal(member);
                 break;
             case KAKAO:
-                memberService.handleLeave(member);
+                handleWithdrawal(member);
                 break;
-//            case NAVER:
-//                memberService.handleLeave(member, authorization);
-//                break;
+            /*
+            case NAVER:
+                handleWithdrawal(member);
+                break;
+             */
             default:
                 throw new CustomException(ErrorCode.MALFORMED_TOKEN);
         }
     }
 
-    public JwtToken reissueAccessToken(JwtTokenReissueDto jwtTokenReissueDto) {
-        return memberService.reissueAccessToken(jwtTokenReissueDto);
+    // 탈퇴 처리
+    private void handleWithdrawal(final Member member) {
+        logout(member);
+        memberService.setMemberStatusLeave(member);
+        log.info("Member {} - LEAVE", member.getEmail());
     }
 
-    public AppleLoginResponse getAppleUserInfo(AppleLoginDto appleLoginDto) {
-        // identityToken 파싱 작업 수행
-        Map<String, String> header = appleTokenParser.parseHeader(appleLoginDto.getIdentityToken());
-
-        // Apple 인증 서버로부터 공개 키 불러옴
-        ApplePublicKeyResponse publicKeys = getApplePublicKeys();
-
-        // 공개 키 중에서 identityToken의 kid, alg 값과 일치하는 키 탐색 후 새로운 공개 키 생성
-        PublicKey publicKey = applePublicKeyGenerator.generate(header, publicKeys);
-
-        // identityToken 검증 작업 수행 및 클레임 추출
-        Claims claims = validateIdentityToken(appleLoginDto.getIdentityToken(), publicKey);
-
-        // Apple RefreshToken 요청
-        String appleRefreshToken = getAppleRefreshToken(appleLoginDto.getAuthorizationCode());
-
-        // 로그인 프로세스 수행 ? 이미 가입된 회원일 경우 JWT 토큰 반환 : 초기 가입 회원일 경우 추가 정보 요청
-        AppleLoginResponse appleLoginResponse = memberService.handleAppleLogin(claims);
-
-        // Apple RefreshToken 저장
-        saveOrUpdateRefreshToken(claims.get("sub", String.class), appleRefreshToken);
-
-        // JWT 토큰 반환
-        return appleLoginResponse;
+    // 애플 로그인 사용자 탈퇴 처리
+    private void handleAppleLoginWithdrawal(final Member member) {
+        socialPlatformService.revokeAppleRefreshToken(member);
+        handleWithdrawal(member);
     }
 
-    public JwtToken continueSignup(AppleSignupContinueDto appleSignupContinueDto) {
-        return memberService.continueAppleSignup(appleSignupContinueDto);
+    // 토큰 Rotation
+    public JwtToken reissueAccessToken(final JwtTokenReissueDto jwtTokenReissueDto) {
+        Claims claims = jwtService.parseClaims(jwtTokenReissueDto.getRefreshToken());
+        return jwtService.reissueAccessToken(memberService.findByEmail(claims.getIssuer()));
     }
 
-    private Claims validateIdentityToken(String identityToken, PublicKey publicKey) {
-        // 클레임 추출
-        Claims claims = appleTokenParser.extractClaims(identityToken, publicKey);
-
-        // iss : 토큰 발행자 검증
-        if (!appleAuthUrl.equals(claims.getIssuer()))
-            throw new CustomException(ErrorCode.TOKEN_ISSUER_MISMATCH);
-
-        // aud : 토큰 수신자 검증
-        if (!appleClientId.equals(claims.getAudience()))
-            throw new CustomException(ErrorCode.TOKEN_AUDIENCE_MISMATCH);
-
-        // exp : 토큰 만료 여부 검증
-        if (claims.getExpiration().before(new Date()))
-            throw new CustomException(ErrorCode.EXPIRED_TOKEN);
-
-        return claims;
-    }
-
-    private ApplePublicKeyResponse getApplePublicKeys() {
-        WebClient webClient = WebClient.builder()
-                .baseUrl(appleAuthUrl)
-                .build();
-
-        return webClient.get()
-                .uri("/auth/keys")
-                .retrieve()
-                .bodyToMono(ApplePublicKeyResponse.class)
-                .block();
-    }
-
-    // Apple 인증 서버에 RefreshToken 요청
-    private String getAppleRefreshToken(String authorizationCode) {
-        MultiValueMap<String, String> body = getCreateTokenBody(authorizationCode);
-
-        WebClient webClient = WebClient.builder()
-                .baseUrl(appleAuthUrl)
-                .build();
-
-        try {
-            AppleTokenResponseDto appleTokenResponse = webClient.post()
-                    .uri("/auth/token")
-                    .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-                    .body(BodyInserters.fromFormData(body))
-                    .retrieve()
-                    .bodyToMono(AppleTokenResponseDto.class)
-                    .block();
-
-            return Objects.requireNonNull(appleTokenResponse).getRefresh_token();
-        } catch (WebClientResponseException e) {
-            throw new CustomException(ErrorCode.APPLE_AUTHORIZATION_CODE_EXPIRED);
-        }
-    }
-
-    private MultiValueMap<String, String> getCreateTokenBody(String authorizationCode) {
-        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-        body.add("code", authorizationCode);
-        body.add("client_id", appleClientId);
-        body.add("client_secret", appleKeyGenerator.generateClientSecret());
-        body.add("grant_type", "authorization_code");
-        return body;
-    }
-
-    public void saveOrUpdateRefreshToken(String socialId, String newRefreshToken) {
-        AppleRefreshToken existingToken = appleRefreshTokenRepository.findBySocialId(socialId).orElse(null);
-
-        // 이미 해당 사용자에 대한 RefreshToken이 존재하는 경우 업데이트
-        if (existingToken != null) {
-            existingToken.updateRefreshToken(newRefreshToken);
-        } else { // 혹은 새로운 RefreshToken 저장
-            AppleRefreshToken newToken = AppleRefreshToken.builder()
-                    .socialId(socialId)
-                    .refreshToken(newRefreshToken)
-                    .build();
-            appleRefreshTokenRepository.save(newToken);
-        }
-    }
-
-    // 애플 로그인 사용자 탈퇴 메소드
-    public void revokeToken(Member member) {
-        AppleRefreshToken refresh = appleRefreshTokenRepository.findBySocialId(member.getSocialId())
-                .orElseThrow(() -> new CustomException(ErrorCode.MALFORMED_TOKEN));
-
-        MultiValueMap<String, String> revokeTokenBody = getRevokeTokenBody(refresh.getRefreshToken());
-
-        WebClient webClient = WebClient.builder()
-                .baseUrl(appleAuthUrl)
-                .build();
-
-        // RefreshToken revoke 요청
-        webClient.post()
-                .uri("/auth/revoke")
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-                .body(BodyInserters.fromFormData(revokeTokenBody))
-                .retrieve()
-                .bodyToMono(Void.class)
-                .block();
-
-        // DB 속 해당 사용자의 RefreshToken 삭제
-        appleRefreshTokenRepository.deleteBySocialId(member.getSocialId());
-
-        // 사용자 상태를 탈퇴로 변경
-        memberService.handleLeave(member);
-
-        log.info("Member {} Account has been successfully revoked.", member.getEmail());
-    }
-
-    public MultiValueMap<String, String> getRevokeTokenBody(String refreshToken) {
-        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-        body.add("client_id", appleClientId);
-        body.add("client_secret", appleKeyGenerator.generateClientSecret());
-        body.add("token", refreshToken);
-        body.add("token_type_hint", "refresh_token");
-        return body;
-    }
 }
